Refactor code structure for improved readability and maintainability
Refactors HTML metadata and improves code readability

Updates the `<head>` section of the HTML to reflect a more personalized project identity, including changes to the title, description, and Open Graph properties.

Adjusts component structure in the Hero component for better maintainability and clarity.

Removes commented-out code for a cleaner look and enhances the user interface with a new favicon.